+++ Tests/NukeUITests/ImageViewExtensionsProgressiveDecodingTests.swift
@@ -5,7 +5,7 @@
 import Testing
 import Foundation
 @testable import Nuke
-@testable import NukeExtensions
+@testable import NukeUI
 
 @Suite(.timeLimit(.minutes(5)))
 struct ImagePipelineProgressiveDecodingTests {
@@ -56,7 +56,7 @@ struct ImagePipelineProgressiveDecodingTests {
         let expectedFinalLoaded = TestExpectation()
 
         // When/Then
-        NukeExtensions.loadImage(
+        NukeUI.loadImage(
             with: Test.request,
             options: options,
             into: imageView,
@@ -94,7 +94,7 @@ struct ImagePipelineProgressiveDecodingTests {
         let expectedFinalLoaded = TestExpectation()
 
         // When/Then
-        NukeExtensions.loadImage(
+        NukeUI.loadImage(
             with: Test.request,
             options: options,
             into: imageView,